Move discrete log into main.
